Cakefile cleanedup

+++ Cakefile
@@ -26,6 +26,7 @@ catch err
     console.log 'WARNING: the which module is required for windows\ntry: npm install which'
   which = null
 
+
 # ANSI Terminal Colors
 bold = '\x1b[0;1m'
 green = '\x1b[0;32m'
@@ -33,7 +34,7 @@ blue = '\x1b[1;36m'
 reset = '\x1b[0m'
 red = '\x1b[0;31m'
 
-task 'docs', 'generate documentation', ->
+task 'doc', 'generate documentation', ->
   docco -> log "docs complete", green
 
 task 'build', 'compile source', ->
@@ -46,64 +47,62 @@ task 'test', 'run tests', ->
   build -> mocha -> log "tests complete", green
 
 task 'dist', 'create distribution', ->
-  #clean -> build -> mocha -> docco -> dist -> log "dist created", green
-  clean -> build -> mocha -> dist -> log "dist created", green
+  # TODO: include docs
+  clean -> build -> mocha -> docco -> dist -> log "dist created", green
 
 task 'clean', 'clean generated files', ->
   clean -> log "clean complete", green
 
-dist = (callback) ->
-    fs     = require("fs")
-    async = require "async"
-    wrench = require('wrench')
+task 'all', 'fresh install', ->
+  # TODO: include docs
+  clean -> install -> build -> mocha -> docco -> log "fresh install complete", green
 
-    read = (filename) ->
-      fs.readFileSync "#{__dirname}/#{filename}", "utf-8"
 
-    conf = JSON.parse(read "package.json")
-
-    name = conf.name
-    version = conf.version
-    tarname= "#{name}-#{version}"
-    tarball= "#{tarname}.tar.gz"
-
-    wrench.rmdirSyncRecursive "./#{tarname}", true
-    fs.mkdirSync "./#{tarname}"
-
-    async.forEach distFiles, (file, cb) ->
-      launch "cp", ["-R", file, "./#{tarname}"], cb
-    , (error) ->
-      if error
-        log error, red
-        fs.rmdirSync "./#{tarname}"
-        return callback()
-      
-      launch "tar", ["cvfz", "./dist/#{tarball}", "./#{tarname}"], (err) ->
-        wrench.rmdirSyncRecursive "./#{tarname}", true
-        return callback()
-
-walk = (dir) ->
-  results = []
-  fs.readdirSync dir, (err, list) ->
-    return [] if err
-    pending = list.length
-    return results unless pending
-    for name in list
-      file = "#{dir}/#{name}"
-      try
-        stat = fs.statSync file
-      catch err
-        stat = null
-      if stat?.isDirectory()
-        _results = walk file
-        results.push name for name in _results
-        return results unless --pending
-      else
-        results.push file
-        return results unless --pending
-
+moduleExists = (name) ->
+  try 
+    require name 
+  catch err 
+    log "#{name} required: npm install #{name}", red
+    false
 
 log = (message, color, explanation) -> console.log color + message + reset + ' ' + (explanation or '')
+
+
+install = (callback) ->
+  log 'installing dependencies', blue
+
+  launch "npm", ["install"], callback
+
+
+dist = (callback) ->
+  log 'creating dist', blue
+
+  async   = moduleExists "async"
+  wrench  = moduleExists "wrench"
+
+  return unless async and wrench
+
+  conf = JSON.parse fs.readFileSync "#{__dirname}/package.json", "utf-8"
+
+  name = conf.name
+  version = conf.version
+  tarname= "#{name}-#{version}"
+  tarball= "#{tarname}.tar.gz"
+
+  wrench.rmdirSyncRecursive "./#{tarname}", true
+  fs.mkdirSync "./#{tarname}"
+
+  async.forEach distFiles, (file, cb) ->
+    launch "cp", ["-R", file, "./#{tarname}"], cb
+  , (error) ->
+    if error
+      log error, red
+      fs.rmdirSync "./#{tarname}"
+      return callback()
+    
+    launch "tar", ["cvfz", "./dist/#{tarball}", "./#{tarname}"], (err) ->
+      wrench.rmdirSyncRecursive "./#{tarname}", true
+      return callback()
 
 launch = (cmd, options=[], callback=(()->)) ->
   cmd = which(cmd) if which
@@ -114,6 +113,7 @@ launch = (cmd, options=[], callback=(()->)) ->
 
 build = (watch, callback) ->
   log 'building', blue
+
   if typeof watch is 'function'
     callback = watch
     watch = false
@@ -125,19 +125,14 @@ build = (watch, callback) ->
 
 clean = (callback) ->
   log 'cleaning', blue
-  launch 'rm', ['-rf', 'lib'], () -> callback?()
-
-moduleExists = (name) ->
-  try 
-    require name 
-  catch err 
-    log "#{name} required: npm install #{name}", red
-    false
+  launch 'rm', ['-rf', 'lib', 'docs'], () -> callback?()
 
 
 mocha = (options, callback) ->
   log 'testing', blue
-  #if moduleExists('mocha')
+
+  return unless moduleExists('mocha')
+
   if typeof options is 'function'
     callback = options
     options = []
@@ -147,10 +142,31 @@ mocha = (options, callback) ->
   
   launch 'mocha', options, callback
 
+
+walk = (dir, done) ->
+  results = []
+  fs.readdir dir, (err, list) ->
+    return done(err, []) if err
+    pending = list.length
+    return done(null, results) unless pending
+    for name in list
+      file = "#{dir}/#{name}"
+      try
+        stat = fs.statSync file
+      catch err
+        stat = null
+      if stat?.isDirectory()
+        walk file, (err, res) ->
+          results.push name for name in res
+          done(null, results) unless --pending
+      else
+        results.push file
+        done(null, results) unless --pending
+
+
 docco = (callback) ->
   log 'docing', blue
-  #if moduleExists('docco')
-  files = walk './src'
-  #files = require("wrench").readdirSyncRecursive 'src'
-  console.log files
-  launch 'docco', files, callback
+  return unless moduleExists('docco')
+  walk "src", (error, files) ->
+    if error then return log "Error reading src: #{err}", red
+    launch 'docco', files, callback
